+++ Blocks/Block Implementations/FullBlock.cs
@@ -1,6 +1,7 @@
 ﻿using OpenTK.Mathematics;
 using OurCraft.Rendering;
 using OurCraft.World;
+using OurCraft.Blocks.Block_Properties;
 
 namespace OurCraft.Blocks.Block_Implementations
 {
@@ -8,15 +9,15 @@ namespace OurCraft.Blocks.Block_Implementations
     public class FullBlock : Block
     {
         //csctr
-        public FullBlock(string name, int bm, int t, int f, int b, int r, int l, int id): 
+        public FullBlock(string name, int bm, int t, int f, int b, int r, int l, ushort id): 
         base(name, bm, t, f,  b, r,  l, id)
         { }
 
-        public FullBlock(string name, int t, int id) :
+        public FullBlock(string name, int t, ushort id) :
         base(name, t, t, t, t, t, t, id)
         { }
 
-        public FullBlock(string name, int b, int t, int s, int id) :
+        public FullBlock(string name, int b, int t, int s, ushort id) :
         base(name, b, t, s, s, s, s, id)
         { }
