+++ Blocks/Block Implementations/GlassBlock.cs
@@ -1,17 +1,17 @@
 ﻿using OpenTK.Mathematics;
 using OurCraft.Rendering;
-using OurCraft.World;
+using OurCraft.Blocks.Block_Properties;
 
 namespace OurCraft.Blocks.Block_Implementations
 {
     //fully transparent glass block, no blending
     public class GlassBlock : FullBlock
     {
-        public GlassBlock(string name, int bm, int t, int f, int b, int r, int l, int id) :
+        public GlassBlock(string name, int bm, int t, int f, int b, int r, int l, ushort id) :
         base(name, bm, t, f, b, r, l, id)
         { }
 
-        public GlassBlock(string name, int t, int id) :
+        public GlassBlock(string name, int t, ushort id) :
         base(name, t, t, t, t, t, t, id)
         { }
